KTOR-462 Move dnsResolver integration tests to JVM-only
The three dnsResolver integration tests use TCP server bind via
withServerSocket and one of them sets pipelining=true; both rely on
features that are not supported on web/native CIO. Keep them in
jvm/test so the common targets stay green.
