+++ ktor-client/ktor-client-cio/common/test/CIOEngineTest.kt
@@ -201,91 +201,6 @@ class CIOEngineTest : ClientEngineTest<CIOEngineConfig>(CIO) {
     }
 
     @Test
-    fun testDnsResolverIsInvokedAndUsedForConnection() = testClient {
-        val resolvedHosts = mutableListOf<String>()
-        config {
-            engine {
-                dnsResolver = { hostname ->
-                    resolvedHosts += hostname
-                    listOf(TEST_SERVER_SOCKET_HOST)
-                }
-            }
-        }
-
-        withServerSocket { client, socket ->
-            launch {
-                val serverPort = (socket.localAddress as InetSocketAddress).port
-                client.get { url(host = "example.invalid", port = serverPort, path = "/") }
-            }
-
-            socket.accept().use {
-                val readChannel = it.openReadChannel()
-                val writeChannel = it.openWriteChannel()
-                readAvailableLines(readChannel)
-                writeOkResponse(writeChannel)
-            }
-
-            assertEquals(listOf("example.invalid"), resolvedHosts)
-        }
-    }
-
-    @Test
-    fun testDnsResolverEmptyResultFailsConnection() = testClient {
-        config {
-            engine {
-                dnsResolver = { emptyList() }
-            }
-        }
-
-        test { client ->
-            assertFailsWith<FailToConnectException> {
-                client.get { url(host = "example.invalid", port = 0, path = "/") }
-            }
-        }
-    }
-
-    @Test
-    fun testDnsResolverThrowingDoesNotLeakConnectionSlot() = testClient {
-        var attempts = 0
-        config {
-            engine {
-                pipelining = true
-                endpoint { maxConnectionsPerRoute = 1 }
-                dnsResolver = {
-                    attempts++
-                    if (attempts == 1) error("simulated DNS failure")
-                    listOf(TEST_SERVER_SOCKET_HOST)
-                }
-            }
-        }
-
-        withServerSocket { client, socket ->
-            val serverPort = (socket.localAddress as InetSocketAddress).port
-
-            val firstFailure = assertFails {
-                client.get { url(host = "example.invalid", port = serverPort, path = "/") }
-            }
-            assertContains(firstFailure.message ?: "", "simulated DNS failure")
-
-            // With pipelining enabled and maxConnectionsPerRoute=1, a leaked counter from the
-            // failed first request would block this second request forever in deliveryPoint.send,
-            // because no pipeline was ever created and the per-route slot is still held.
-            launch {
-                socket.accept().use {
-                    val readChannel = it.openReadChannel()
-                    val writeChannel = it.openWriteChannel()
-                    readAvailableLines(readChannel)
-                    writeOkResponse(writeChannel)
-                }
-            }
-            val response = withTimeout(5.seconds) {
-                client.get { url(host = "example.invalid", port = serverPort, path = "/") }
-            }
-            assertEquals(HttpStatusCode.OK, response.status)
-        }
-    }
-
-    @Test
     fun `test ignore invalid Content-Length with Connection close`() = runTest {
         val serverResponse = getResponse(connection = "close")
         val response = readResponse(GMTDate.START, anyRequest, serverResponse, ByteChannel(), coroutineContext)
